backend/requirements.txt
psycopg2-binary>=2.9,<3
Flask>=2.2,<3
python-dotenv>=1.0,<2
